api/requirements.txt
telethon==1.36.0
asyncpg==0.30.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6